+++ R/read.R
@@ -98,12 +98,17 @@ oe_read = function(
   check_layer_provider(layer, provider)
 
   # Check that all arguments inside ... are named arguments. See also
-  # https://github.com/ropensci/osmextract/issues/234
-  if (...length() && any(is.na(...names()))) {
-    stop(
-      "All arguments in oe_get and oe_read beside 'place' and 'layer' must be named. ",
-      "Please check also that you didn't add an extra comma at the end of your call.",
-      call. = FALSE
+  # https://github.com/ropensci/osmextract/issues/234. I extract the names in
+  # ... and save the result in dots_names since names(list(...)) returns an
+  # error when there is a missing element in ... See the examples in utils.R. I
+  # need to check null and "" values. See utils.R for examples. See
+  # https://github.com/ropensci/osmextract/issues/241 and corresponding PR for a
+  # discussion.
+  dots_names = extract_dots_names_safely(...)
+  if (...length() && (any(is.null(dots_names)) | any(dots_names == ""))) {
+    stop_custom(
+      .subclass = "osmext-names-dots-error",
+      message = "All arguments in oe_get() and oe_read() beside 'place' and 'layer' must be named. Please check also that you didn't add an extra comma at the end of your call.",
     )
   }
 
@@ -111,7 +116,7 @@ oe_read = function(
   # https://github.com/ropensci/osmextract/issues/122. Moreover, I had to use
   # ...names() instead of names(list(...)) because of
   # https://github.com/ropensci/osmextract/issues/234
-  if ("query" %in% ...names()) {
+  if ("query" %in% dots_names) {
     # Check if the query argument (which is passed to sf::st_read) was defined using a
     # layer different than layer argument. Indeed:
     # Extracted from sf::st_read docs: For query with a character dsn the query
@@ -170,7 +175,7 @@ oe_read = function(
     # returns a notes in the R CMD checks related to :::. Hence, I decided to
     # use names(formals("st_read.character", envir = getNamespace("sf")))
     any(
-      ...names() %!in%
+      dots_names %!in%
       # The ... arguments in st_read are passed to st_as_sf so I need to add the
       # formals of st_as_sf.
       # See https://github.com/ropensci/osmextract/issues/152
@@ -185,7 +190,7 @@ oe_read = function(
       "The following arguments were probably misspelled: ",
       paste(
         setdiff(
-          ...names(),
+          dots_names,
           union(
             names(formals(get("st_read.character", envir = getNamespace("sf")))),
             names(formals(get("st_as_sf.data.frame", envir = getNamespace("sf"))))